// React js/persimmon-app-main/src/app/types/customization.ts
export interface CustomizationSettings {
    coverImage: string;
    heading: string;
    description: string;
    darkMode: boolean;
    primaryColor: string;
    fontStyle: string;
    allPrimaryColors: string[];
    showCover: boolean;
    icon: string | null
    iconLink: string
    headerColor: string;
    headerColors:string[];
    careerPageUrl:string;
  }